Kyoraku duel finish, DS, Ores

+++ UHC/src/fr/lastril/uhchost/modes/bleach/commands/CmdTempete.java
@@ -5,6 +5,7 @@ import fr.lastril.uhchost.enums.Messages;
 import fr.lastril.uhchost.modes.bleach.roles.shinigamis.soulsociety.Byakuya;
 import fr.lastril.uhchost.modes.command.ModeSubCommand;
 import fr.lastril.uhchost.player.PlayerManager;
+import fr.lastril.uhchost.player.modemanager.BleachPlayerManager;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -35,14 +36,19 @@ public class CmdTempete implements ModeSubCommand {
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         Player player = (Player) sender;
         PlayerManager playerManager = main.getPlayerManager(player.getUniqueId());
+        BleachPlayerManager bleachPlayerManager = playerManager.getBleachPlayerManager();
         if (!playerManager.hasRole() || !playerManager.isAlive()) {
             return false;
         }
         if(playerManager.getRole() instanceof Byakuya){
             Byakuya byakuya = (Byakuya) playerManager.getRole();
-            if(byakuya.hasUsedZone()){
-                byakuya.deleteZone();
-                player.sendMessage("§cVous venez de détruire votre zone ! Les joueurs ont été ejectés de celle-ci.");
+            if(bleachPlayerManager.canUsePower()){
+                if(byakuya.hasUsedZone()){
+                    byakuya.deleteZone();
+                    player.sendMessage("§cVous venez de détruire votre zone ! Les joueurs ont été ejectés de celle-ci.");
+                }
+            } else {
+                player.sendMessage(Messages.BLEACH_PREFIX.getMessage() + Messages.CANT_USE_POWER_NOW.getMessage());
             }
         }
         return false;
